backend-api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1

